# src/api/swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "Zombie crud",
      "version": "1.0.0",
      "title": "Zombie crud",
      "termsOfService": "http://swagger.io/terms/",
      "contact": {
        "email": "apiteam@swagger.io"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      }
    },
    "host": "zombiestore.swagger.io",
    "basePath": "/v2",
    "tags": [
      {
        "name": "zombie",
        "description": "Everything about your Zombies",
        "externalDocs": {
          "description": "Find out more",
          "url": "http://swagger.io"
        }
      }
    ],
    "schemes": [
      "https",
      "http"
    ],
    "paths": {
      "/zombies": {
        "post": {
          "tags": [
            "zombie"
          ],
          "summary": "Add a new zombie",
          "description": "",
          "operationId": "addZombie",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Zombie object that needs to be added to the store",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Zombie"
              }
            }
          ],
          "responses": {
            "405": {
              "description": "Invalid input"
            }
          }
        }
      },
      "/zombies/{zombieId}": {
        "get": {
          "tags": [
            "zombie"
          ],
          "summary": "Find zombie by ID",
          "description": "Returns a single zombie",
          "operationId": "getZombieById",
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [
            {
              "name": "zombieId",
              "in": "path",
              "description": "ID of zombie to return",
              "required": true,
              "type": "integer",
              "format": "int64"
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "$ref": "#/definitions/Zombie"
              }
            },
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Zombie not found"
            }
          }
        },
        "delete": {
          "tags": [
            "zombie"
          ],
          "summary": "Deletes a zombie",
          "description": "",
          "operationId": "deleteZombie",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "api_key",
              "in": "header",
              "required": false,
              "type": "string"
            },
            {
              "name": "zombieId",
              "in": "path",
              "description": "Zombie id to delete",
              "required": true,
              "type": "integer",
              "format": "int64"
            }
          ],
          "responses": {
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Zombie not found"
            }
          }
        },
        "put": {
          "tags": [
            "zombie"
          ],
          "summary": "Update an existing zombie",
          "description": "",
          "operationId": "updateZombie",
          "consumes": [
            "application/json",
            "application/xml"
          ],
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Zombie object that needs to be added to the store",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Zombie"
              }
            },
            {
              "name": "zombieId",
              "in": "path",
              "description": "Zombie id to update",
              "required": true,
              "type": "integer",
              "format": "int64"
            }
          ],
          "responses": {
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Zombie not found"
            },
            "405": {
              "description": "Validation exception"
            }
          }
        }
      }
    },
    "definitions": {
      "Item": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "integer",
            "format": "int64"
          }
        },
        "xml": {
          "name": "Item"
        }
      },
      "Zombie": {
        "type": "object",
        "required": [
          "name",
          "photoUrls"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "example": "John"
          },
          "items": {
            "type": "array",
            "xml": {
              "name": "tag",
              "wrapped": true
            },
            "items": {
              "$ref": "#/definitions/Item"
            }
          }
        },
        "xml": {
          "name": "Zombie"
        }
      },
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    },
    "externalDocs": {
      "description": "Find out more about Swagger",
      "url": "http://swagger.io"
    }
  }